# json: improve performance by using a pool of scanners